requirements.txt
openai
python-dotenv
pyautogui
Pillow
